fix: use caching for pytest only on main (#7)
Ensure that only the main branch caches the test data, otherwise
PR-specific test data caches will evict the main branch test data cache
and cause cache misses more often. This is because everytime a new PR is
created that branch will not use the main data cache since the caches
are branch-specific and create a new cache, evicting the old cache since
the data is ~9GB and Github only allows a total cache size of 10GB.

Instead, we might as well not cache for PRs / other branches and only
cache for the main branch, so we still have some use for the caching and
don't constantly evict the cached test data. If we try to keep caching
for all branches, we'll end up not actually getting cache hits very
often because of the repo-wide 10GB Github limit, caches themselves
being limited to only their specific branch, and the size of our test
dataset.